pyproject.toml
[tool.poetry]
name = "tomoro_rag"
version = "0.1.0"
description = "RAG system for financial documents"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
openai = "^1.0.0"
qdrant-client = "^1.7.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.2"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 